requirements.txt
torch
faster-whisper
google-generativeai
speechrecognition
gtts
pydub
fastapi
uvicorn
python-dotenv

